requirements.txt
# TiPAI-TSPO POC Requirements
# Python 3.10+

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Image Generation (SDXL + FLUX dependencies)
diffusers>=0.24.0
accelerate>=0.25.0
transformers>=4.35.0
safetensors>=0.4.0
invisible-watermark>=0.2.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# CLIP Scoring
open-clip-torch>=2.23.0

# Segmentation (SAM)
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data & Visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0

# Demo UI
gradio>=4.0.0

# Utilities
tqdm>=4.65.0

# HuggingFace Dataset
datasets>=2.14.0
huggingface-hub>=0.19.0
